+++ server/zanata-war/src/main/java/org/zanata/config/ObsoleteJNDIChecker.java
@@ -18,7 +18,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.zanata.migration;
+package org.zanata.config;
 
 import java.util.Set;
 import javax.naming.InitialContext;
